glof/requirements.txt
numpy==1.24.3
pandas==2.0.3
tensorflow==2.13.0
opencv-python==4.8.0.76
scikit-learn==1.3.0
pillow==10.0.0
fastapi==0.103.1
uvicorn==0.23.2
python-multipart==0.0.6
python-jose==3.3.0
passlib==1.7.4
python-dotenv==1.0.0
requests==2.31.0
pydantic==2.3.0 